rename transformers and update test
